pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "adversarial-queueing"
version = "0.1.0"
description = "Simulation experiments for adversarial queueing/control Markov games."
requires-python = ">=3.11"
dependencies = [
  "numpy",
  "scipy",
  "PyYAML",
]

[project.optional-dependencies]
dev = [
  "pytest",
]

[tool.setuptools.packages.find]
where = ["src"]

